+++ Samples/2.0/Common/src/System/Desktop/MainLoopMultiThreaded.cpp
@@ -130,7 +130,7 @@ unsigned long renderThreadApp( Ogre::ThreadHandle *threadHandle )
 
     Ogre::Timer timer;
 
-    unsigned long startTime = timer.getMicroseconds();
+    unsigned long long startTime = timer.getMicroseconds();
 
     double timeSinceLast = 1.0 / 60.0;
 
@@ -146,7 +146,7 @@ unsigned long renderThreadApp( Ogre::ThreadHandle *threadHandle )
             Ogre::Threads::Sleep( 500 );
         }
 
-        unsigned long endTime = timer.getMicroseconds();
+        unsigned long long endTime = timer.getMicroseconds();
         timeSinceLast = (endTime - startTime) / 1000000.0;
         timeSinceLast = std::min( 1.0, timeSinceLast ); //Prevent from going haywire.
         startTime = endTime;
